+++ qwt/src/qwt_plot_histogram.cpp
@@ -341,7 +341,7 @@ void QwtPlotHistogram::drawSeries( QPainter* painter,
         return;
 
     if ( to < 0 )
-        to = dataSize() - 1;
+        to = static_cast<int>( dataSize() - 1 );
 
     switch ( m_data->style )
     {
