+++ qwt/src/qwt_plot_vectorfield.cpp
@@ -812,7 +812,7 @@ void QwtPlotVectorField::drawSeries( QPainter* painter,
         return;
 
     if ( to < 0 )
-        to = dataSize() - 1;
+        to = static_cast<int>( dataSize() - 1 );
 
     if ( from < 0 )
         from = 0;
